+++ mnist.h
@@ -9,4 +9,4 @@ using namespace std;
 
 void read_Mnist(string filename, vector<vector<float> > &vec);
 
-void read_Mnist_Label(string filename, vector<float> &vec);
+void read_Mnist_Label(string filename, vector<int> &vec);
